refactor(frontend): 読み上げ・札めくり進行をuseKarutaReadingへ切り出す（issue #607）
設計:
- issue #607（App.jsx/App.test.jsxの肥大化解消）の残り抽出単位のうち
  最後の1項目、読み上げ・札めくり進行（playKarutaInternal、タイマー、
  フェード演出、結果確定）をhooks/useKarutaReading.jsへ切り出した
- この項目は、同じコード領域の未解決バグissue #729（自動読み上げの
  途中停止）と並行させるとリファクタリング起因の不具合と既存バグの
  切り分けが難しくなるため、#729の解決を待って着手する方針だった
  （issue #607コメント参照）。#729は自動読み上げモード自体の廃止で
  解決したため、今回着手した
- 切り出した状態・ref: currentPhrase/displayContent/isFadingOut/
  audioQueue/isReading/loading/isAllRead等のstateと、flipTimeoutRef等の
  内部ref一式
- 切り出した関数: playAudio/playIntroSound/playCongratulationAudio/
  revealCurrentResult/playKarutaInternal/playKaruta/stopReading、
  および結果確定後のフェード遷移・次札プリフェッチのuseEffect
- App.jsx側は、hookが返すqueueRepeatAudio(audioData)を使う薄い
  repeatPhraseラッパーのみ残した（詳細画面の対象取得はApp.jsx側の
  責務のため）
- resetGame/restartCategoryが個別に行っていたcurrentPhrase/
  displayContent/isAllRead/isFadingOutのリセット処理が完全に同一
  だったため、hookが公開するresetReadingState()に統合し重複を解消した
- 却下内容: App.test.jsxの対応するテスト分割は見送った。読み上げ・
  札めくりの挙動は画面遷移やボタンの活性状態と一体で検証する必要があり
  （PR #622のuseQuizRoomAdminのような独立したAPIサーフェスと異なる）、
  renderHook単位での再現は同等のUIモックを重複して持つだけで得るものが
  薄いと判断した

影響:
- frontend/src/App.jsx: 1,709行→1,301行
- frontend/src/hooks/useKarutaReading.js: 新規487行
- 読み上げ・札めくりの挙動（音声再生順序、フェード演出タイミング、
  結果確定、issue #729由来のaudio.play()タイムアウト保護、issue #721の
  最終札での「次の札」無効化等）に機能的な変更はない

テスト:
- frontend: npx eslint . エラー0件
- frontend: npx vitest run src/App.test.jsx（単独実行）で65件全て成功
- backend: npx eslint . エラー0件（今回未変更だが対象パッケージとして確認）

issue #607は継続対応（App.jsxがクローズ基準の1,000行を上回るため引き続きオープン）
